requirement.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-env==0.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
